cleaning and restructure the packages and rename the source file
