+++ TechnicalSepc.rst
@@ -7,7 +7,7 @@ You need to install tox to run the workflow tox task or env run task.
 
 .. code-block:: shell
 
-   python -m pip install tox 
+   python -m pip install tox
    tox -v -s false -e pin-deps
    tox -e type -v -s false
    tox -v -s false | tee test_output.log
@@ -19,7 +19,7 @@ You need to install tox to run the workflow tox task or env run task.
    tox -e ruff -vv -s false
    tox -e prospector -vv -s false
 
-So before pushing to the remote repository run the above commands one by one and check for errors and fix them. 
+So before pushing to the remote repository run the above commands one by one and check for errors and fix them.
 After all are correct with no errors then you push to the remote repo on github.
 
 THIS SECTION FOR ISO
@@ -29,31 +29,32 @@ One language Text To Many languages Text Translation:
 -----------------------------------------------------
 
 | Your backend library ( multimedia-translator) does not know anything about the rest . The library is rest agonistic ( does not have any clue)
-So somebody will call your library  with a payload called "TextTransModel" 
-Your library  will have its own implementation to translate into different languages ( LlibreTranslate local server ) . You Library needs to make sure the language are supported ( your library needs to adhere to the LiberateTranslate languages supported since your library has no capability to do endless languages but rather follows the LilbreTranslate)
-Your library needs to validate the model (TextTransModel) to make sure it can do the job. If one of the selected languages (targets or source ) is not supported, you need to raise an exception to explain this . The caller ( the web application which uses your library) will handle the exception . 
-If the validation is green, your library will communicate with the server ( it is your implementation hidden to me as a caller. I don't care about your internal implementation ) and return back a python model  see below correction for the model (TextModel) 
-
-Update the spec to explain the above in your branch as usual . Let me know . You can push the json object down to my section since I need them later 
+  So somebody will call your library  with a payload called "TextTransModel"
+  Your library  will have its own implementation to translate into different languages ( LlibreTranslate local server ) . You Library needs to make sure the language are supported ( your library needs to adhere to the LiberateTranslate languages supported since your library has no capability to do endless languages but rather follows the LilbreTranslate)
+  Your library needs to validate the model (TextTransModel) to make sure it can do the job. If one of the selected languages (targets or source ) is not supported, you need to raise an exception to explain this . The caller ( the web application which uses your library) will handle the exception .
+  If the validation is green, your library will communicate with the server ( it is your implementation hidden to me as a caller. I don't care about your internal implementation ) and return back a python model  see below correction for the model (TextModel)
 
 
-class  TextModel(): # generic text with language can be used for response 
-    def __init__(self, text:str, lang:LANGUAGES) -> None:
-        self.__title = title
-        self.__lang = lang
+| Update the spec to explain the above in your branch as usual . Let me know . You can push the json object down to my section since I need them later
+
+.. code-block:: shell
+
+  class  TextModel(): # generic text with language can be used for response
+      def __init__(self, text:str, lang:LANGUAGES) -> None:
+          self.__title = title
+          self.__lang = lang
 
 
-while this 
 
-class TextTransModel(TextModel): # this is the one your receive 
+  class TextTransModel(TextModel): # this is the one your receive
 
-    def __init__(self, text:str, lang:LANGUAGES, targets:list[LANGUAGES]) -> None:
-        uper().__init__(title, lang) # common
-        self.__targets = targets  # 
-      
+      def __init__(self, text:str, lang:LANGUAGES, targets:list[LANGUAGES]) -> None:
+          uper().__init__(title, lang) # common
+          self.__targets = targets  #
 
 
-Now for the body, it will be the split into two calls for the above (TextTransModel) since  body and title  can be one if you send them as one text with "|" seperator ( saving two calls into one call) and once you get back the translated "title|body" you can split them anb build the response back ( You do it later since you need to implement the above first and make end to end integration test ( inside the main of the class as we use to do ) 
+
+Now for the body, it will be the split into two calls for the above (TextTransModel) since  body and title  can be one if you send them as one text with "|" seperator ( saving two calls into one call) and once you get back the translated "title|body" you can split them anb build the response back ( You do it later since you need to implement the above first and make end to end integration test ( inside the main of the class as we use to do )
 
 
 | In order to test, you need to innstall the libretranslate server and start it locally see
@@ -66,7 +67,7 @@ Now for the body, it will be the split into two calls for the above (TextTransMo
 | you need to response with a given error json object see model which explain the issue
 
 | You need to avoid using third party liberary if possible to make http call(rest call with json payload)  **use native if you can** https://docs.python.org/3/library/http.client.html
-| 
+|
 
 | Step:
 
@@ -164,8 +165,8 @@ Install pytest and coverage in your env.
 This is the documentation for the coverage: https://coverage.readthedocs.io/en/7.7.0/
 
 The tests should cover about 95% of the code.
-To check how much your tests are covering: 
-$ coverage run -m pytest 
+To check how much your tests are covering:
+$ coverage run -m pytest
 $ coverage report -m
