+++ lab/libretranslatepy/conf/languages.json
@@ -1,3 +1,4 @@
+# Taken from libraTranslate for support languages 
 [
     {
         "code": "en",
