+++ src/multimedia_translator/audio/__init__.py
@@ -1,0 +1,6 @@
+'''
+Package for **Audio Translation**
+
+| One Language to Many Language
+
+'''
