+++ src/multimedia_translator/text/__init__.py
@@ -1,0 +1,6 @@
+'''
+Package for **Text Translation**
+
+| One Language to Many Languages
+
+'''
