+++ src/multimedia_translator/text/sender_dummy.py
