+++ src/multimedia_translator/video/__init__.py
@@ -1,0 +1,6 @@
+'''
+Package for **Video Translation**
+
+| One Language to Many Language
+
+'''
